Microsoft is changing app registration link
As of "May 2019" the link for registering Microsoft Login Applications is changing and the old link will no longer work. If you click on the old link and try to add an application you get a popup telling you the link will stop working soon and that you should use the Azure Portal link to register applications now.